implemented API authentication with laravel sanctum

+++ app/Http/Controllers/Api/ProductsController.php
@@ -6,10 +6,12 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\ProductResource;
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Response;
 
 class ProductsController extends Controller
 {
+
     /**
      * Display a listing of the resource.
      */
@@ -34,6 +36,12 @@ class ProductsController extends Controller
             'price' => 'required|numeric|min:0',
             'compare_price' => 'nullable|numeric|gt:price'
         ]);
+
+        $user = $request->user();
+        if (!$user->tokenCan('products.create')) {
+            abort(403, 'Not allowed');
+        }
+
         $product = Product::create($request->all());
 
         return $product;
@@ -61,6 +69,10 @@ class ProductsController extends Controller
             'price' => 'sometimes|required|numeric|min:0',
             'compare_price' => 'nullable|numeric|gt:price',
         ]);
+        $user = $request->user();
+        if (!$user->tokenCan('products.update')) {
+            abort(403, 'Not allowed');
+        }
         $product->update($request->all());
 
         return Response::json($product);
@@ -71,6 +83,10 @@ class ProductsController extends Controller
      */
     public function destroy(string $id)
     {
+        $user = Auth::guard('sanctum')->user();
+        if (!$user->tokenCan('products.delete')) {
+            return response(['message' => 'Not allowed'], 403);
+        }
         Product::destroy($id);
         return ['message' => 'Product deleted successfully'];
     }
